Moved ace to assets under json-editor
