+++ src/components/json-editor/assets/json-editor/ace/ext-searchbox.js
