+++ src/components/json-editor/test/json-editor.spec.tsx
@@ -16,7 +16,7 @@
 
 import { newSpecPage } from '@stencil/core/testing';
 import { JsonEditor } from '../json-editor';
-import * as translations from '../../../../nls/Elm_strings.json';
+import * as translations from '../../../assets/nls/Elm_strings.json';
 import { h } from '@stencil/core';
 
 describe('json-editor', () => {
